refactor(component): button
