+++ src/components/Button/index.ts
@@ -1,3 +1,2 @@
 export * from './Button';
-export * from './styled';
 export * from './types';
