--- src/components/ButtonNew/index.ts
@@ -1,2 +1,0 @@
-export * from './Button';
-export * from './types';
