--- src/components/ButtonNew/types.ts
@@ -1,4 +1,0 @@
-export const PRIMARY_TYPE = 'primary';
-export const SECONDARY_TYPE = 'secondary';
-
-export type ButtonType = typeof PRIMARY_TYPE | typeof SECONDARY_TYPE;
